Created tests
Have now started with testing.
These are so far pretty basic, 'test_history' will check if history file
exists and also validate it as a clean file.
'test_init' will run the 'init' function inside shell and will hopefully
catch if a builtin command that is being registered is correct or not.